Migrate to Maven project
