+++ .gitignore
@@ -1,3 +1,4 @@
 /nbproject/private/
 /build/
 /dist/
+/target/
