--- lib/CVS/Entries
@@ -1,7 +1,0 @@
-/cpptasks.jar/1.1/Mon Jan 22 18:43:16 2007/-kb/
-/dom.jar/1.3/Mon Sep 25 23:45:44 2006/-kb/
-/jaxp-api.jar/1.3/Mon Sep 25 23:45:44 2006/-kb/
-/sax.jar/1.3/Mon Sep 25 23:45:44 2006/-kb/
-/xalan.jar/1.3/Mon Sep 25 23:45:44 2006/-kb/
-/xercesImpl.jar/1.3/Mon Sep 25 23:45:44 2006/-kb/
-D
