--- lib/CVS/Repository
@@ -1,1 +1,0 @@
-xmlpg/lib
