--- lib/CVS/Root
@@ -1,1 +1,0 @@
-:ext:mcgredo@xmsf.cvs.sourceforge.net:/cvsroot/xmsf
