--- nbproject/genfiles.properties
@@ -1,5 +1,0 @@
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=aa1a40eb
-nbproject/build-impl.xml.script.CRC32=8731263b
-nbproject/build-impl.xml.stylesheet.CRC32=830a3534@1.80.1.48
