--- nbproject/private/private.properties
@@ -1,6 +1,0 @@
-compile.on.save=false
-do.depend=false
-do.jar=true
-javac.debug=true
-javadoc.preview=true
-user.properties.file=/Users/mcgredo/Library/Application Support/NetBeans/8.2/build.properties
