--- nbproject/project.properties
@@ -1,101 +1,0 @@
-annotation.processing.enabled=true
-annotation.processing.enabled.in.editor=false
-annotation.processing.processors.list=
-annotation.processing.run.all.processors=true
-annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
-application.title=xmlpg
-application.vendor=mcgredo
-build.classes.dir=${build.dir}/classes
-build.classes.excludes=**/*.java,**/*.form
-# This directory is removed when the project is cleaned:
-build.dir=build
-build.generated.dir=${build.dir}/generated
-build.generated.sources.dir=${build.dir}/generated-sources
-# Only compile against the classpath explicitly listed here:
-build.sysclasspath=ignore
-build.test.classes.dir=${build.dir}/test/classes
-build.test.results.dir=${build.dir}/test/results
-buildfile=nbbuild.xml
-# Uncomment to specify the preferred debugger connection transport:
-#debug.transport=dt_socket
-debug.classpath=\
-    ${run.classpath}
-debug.test.classpath=\
-    ${run.test.classpath}
-# This directory is removed when the project is cleaned:
-dist.dir=dist
-dist.jar=${dist.dir}/xmlpg.jar
-dist.javadoc.dir=${dist.dir}/javadoc
-endorsed.classpath=
-excludes=
-file.reference.activation.jar=lib/activation.jar
-file.reference.cpptasks.jar=lib/cpptasks.jar
-file.reference.dom.jar=lib/dom.jar
-file.reference.jaxb-api.jar=lib/jaxb-api.jar
-file.reference.jaxb-impl.jar=lib/jaxb-impl.jar
-file.reference.jaxb-xjc.jar=lib/jaxb-xjc.jar
-file.reference.jaxb1-impl.jar=lib/jaxb1-impl.jar
-file.reference.jaxp-api.jar=lib/jaxp-api.jar
-file.reference.jsr173_1.0_api.jar=lib/jsr173_1.0_api.jar
-file.reference.sax.jar=lib/sax.jar
-file.reference.trunk-src=src
-file.reference.xalan.jar=lib/xalan.jar
-file.reference.xercesImpl.jar=lib/xercesImpl.jar
-includes=**
-jar.compress=false
-javac.classpath=\
-    ${file.reference.activation.jar}:\
-    ${file.reference.cpptasks.jar}:\
-    ${file.reference.dom.jar}:\
-    ${file.reference.jaxb-api.jar}:\
-    ${file.reference.jaxb-impl.jar}:\
-    ${file.reference.jaxb-xjc.jar}:\
-    ${file.reference.jaxb1-impl.jar}:\
-    ${file.reference.jaxp-api.jar}:\
-    ${file.reference.jsr173_1.0_api.jar}:\
-    ${file.reference.sax.jar}:\
-    ${file.reference.xalan.jar}:\
-    ${file.reference.xercesImpl.jar}
-# Space-separated list of extra javac options
-javac.compilerargs=
-javac.deprecation=false
-javac.external.vm=false
-javac.processorpath=\
-    ${javac.classpath}
-javac.source=1.8
-javac.target=1.8
-javac.test.classpath=\
-    ${javac.classpath}:\
-    ${build.classes.dir}:\
-    ${libs.junit.classpath}:\
-    ${libs.junit_4.classpath}
-javac.test.processorpath=\
-    ${javac.test.classpath}
-javadoc.additionalparam=
-javadoc.author=false
-javadoc.encoding=${source.encoding}
-javadoc.noindex=false
-javadoc.nonavbar=false
-javadoc.notree=false
-javadoc.private=false
-javadoc.splitindex=true
-javadoc.use=true
-javadoc.version=false
-javadoc.windowtitle=
-main.class=edu.nps.moves.xmlpg.Xmlpg
-manifest.file=manifest.mf
-meta.inf.dir=${src.dir}/META-INF
-mkdist.disabled=false
-platform.active=default_platform
-run.classpath=\
-    ${javac.classpath}:\
-    ${build.classes.dir}
-# Space-separated list of JVM arguments used when running the project
-# (you may also define separate properties like run-sys-prop.name=value instead of -Dname=value
-# or test-sys-prop.name=value to set system properties for unit tests):
-run.jvmargs=
-run.test.classpath=\
-    ${javac.test.classpath}:\
-    ${build.test.classes.dir}
-source.encoding=UTF-8
-src.dir=${file.reference.trunk-src}
