+++ src/main/java/edu/nps/moves/xmlpg/BitField.java
