+++ src/main/java/edu/nps/moves/xmlpg/ClassAttribute.java
