+++ src/main/java/edu/nps/moves/xmlpg/GeneratedClass.java
