+++ src/main/java/edu/nps/moves/xmlpg/InitialValue.java
