+++ src/main/java/edu/nps/moves/xmlpg/JavascriptGenerator.java
